Add exception handling

+++ Backend/gymcms/src/main/java/damcio/gymcms/banner/BannerService.java
@@ -3,6 +3,8 @@ package damcio.gymcms.banner;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import damcio.gymcms.exception.ResourceNotFoundException;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -15,8 +17,12 @@ public class BannerService {
         return bannerRepository.save(banner);
     }
 
-    public Optional<Banner> getBannerById(Integer id){
-        return bannerRepository.findById(id);
+    public Banner getBannerById(Integer id){
+        Optional<Banner> banner = bannerRepository.findById(id);
+        if (banner.isEmpty())
+            throw new ResourceNotFoundException("Banner not found");
+        
+        return banner.get();
     }
 
     public List<Banner> getAllBanners(){
@@ -28,7 +34,11 @@ public class BannerService {
     }
 
     public Banner updateBanner(Banner banner){
-        Banner existingBanner = bannerRepository.findById(banner.getId()).get();
+        Optional<Banner> optionalExistingBanner = bannerRepository.findById(banner.getId());
+        if (optionalExistingBanner.isEmpty())
+            throw new ResourceNotFoundException("Couldn't find banner to update");
+        
+        Banner existingBanner = optionalExistingBanner.get();
         existingBanner.setActive(banner.getActive());
         existingBanner.setBody(banner.getBody());
         existingBanner.setTitle(banner.getTitle());
@@ -36,6 +46,9 @@ public class BannerService {
     }
 
     public void deleteBanner(Integer id){
+        if (bannerRepository.existsById(id))
+            throw new ResourceNotFoundException("Couldn't find banner to change");
+        
         bannerRepository.deleteById(id);
     }
 }
